[KED-2517] Fix focus mode when modular pipelines are contracted
